+++ contrib/10.math/math.c
@@ -1,4 +1,5 @@
 #include "picrin.h"
+#include "picrin/extra.h"
 
 #include <math.h>
