+++ contrib/20.r7rs/src/file.c
@@ -3,6 +3,7 @@
  */
 
 #include "picrin.h"
+#include "picrin/extra.h"
 
 #include <stdio.h>
