+++ contrib/20.r7rs/src/load.c
@@ -3,6 +3,7 @@
  */
 
 #include "picrin.h"
+#include "picrin/extra.h"
 
 #include <stdio.h>
