+++ contrib/20.r7rs/src/mutable-string.c
@@ -1,4 +1,5 @@
 #include "picrin.h"
+#include "picrin/extra.h"
 #include "picrin/object.h"
 
 void
