+++ contrib/30.random/src/random.c
@@ -1,4 +1,5 @@
 #include "picrin.h"
+#include "picrin/extra.h"
 
 double genrand_real3(void);
