+++ contrib/30.readline/src/readline.c
@@ -6,6 +6,7 @@
 forget to use the C++ extern "C" to get it to compile.
 */
 #include "picrin.h"
+#include "picrin/extra.h"
 
 #include <editline/readline.h>
