+++ contrib/30.regexp/src/regexp.c
@@ -1,4 +1,5 @@
 #include "picrin.h"
+#include "picrin/extra.h"
 
 #include <regex.h>
