+++ contrib/60.repl/repl.c
@@ -1,12 +1,11 @@
 #include "picrin.h"
+#include "picrin/extra.h"
 
 #include <unistd.h>
-
 
 static pic_value
 pic_repl_tty_p(pic_state *pic)
 {
-
   pic_get_args(pic, "");
 
   return pic_bool_value(pic, (isatty(STDIN_FILENO)));
