+++ extlib/benz/vector.c
@@ -3,6 +3,7 @@
  */
 
 #include "picrin.h"
+#include "picrin/extra.h"
 #include "picrin/object.h"
 
 pic_value
